16k: Change the userdata partition to ext4
F2FS is broken with 16kB pages. So we change it to ext4 instead.
Removed the f2fs specific options from fstab.
We also need to add a fstab file to system image and change the
fstab sufix because we cannot modify the fstab in the vendor image
which is prebuilt.

Bug: 260640495

Change-Id: I64c4a2510ef1c9c38cb33ba671d36ee888ad46a6